+++ view.cpp
@@ -2,7 +2,7 @@
 #include "consts.h"
 #include "canvas.h"
 
-View::View(double x_left, double x_right, double y_top, double y_bottom)
+View::View(long double x_left, long double x_right, long double y_top, long double y_bottom)
 	: QMainWindow()
 {
